### 2020-09-10 #2660 @whedon
[PRE REVIEW]: J2suscep: Calculation of magnetic exchange coupling and temperature dependence of magnetic susceptibility
**Submitting author:** @Swetanshu (<a href="http://orcid.org/0000-0002-0626-0626">Swetanshu Tandon</a>)
**Repository:** <a href="https://github.com/WatsonGroupTCD/J2suscep" target ="_blank">https://github.com/WatsonGroupTCD/J2suscep</a>
**Version:** 1.0
