+++ JOSS_paper/paper.md
@@ -24,6 +24,8 @@ date: September 2020
 bibliography: paper.bib
 ---
 
+# `Statement of need`
+
 The field of molecular magnetism has attracted significant interest owing 
 to the rising need to miniaturize magnets. Advancement in this field 
 requires a deep understanding of the electronic structure, and computational 
