+++ JOSS_paper/paper.bib
@@ -68,7 +68,7 @@ url="https://doi.org/10.1007/128_2015_645"
 @Article{Neese,
   author   = {Frank Neese},
   title    = {Prediction of molecular properties and molecular spectroscopy with density functional theory: From fundamental theory to exchange-coupling},
-  doi      = {https://doi.org/10.1016/j.ccr.2008.05.014},
+  doi      = {10.1016/j.ccr.2008.05.014},
   issn     = {0010-8545},
   note     = {Theory and Computing in Contemporary Coordination Chemistry},
   number   = {5},
@@ -98,7 +98,7 @@ url="https://doi.org/10.1007/128_2015_645"
 @Article{magpack,
   author   = {E. Ramos and J.E. Roman and S. Cardona-Serra and J.M. Clemente-Juan},
   title    = {Parallel implementation of the MAGPACK package for the analysis of high-nuclearity spin clusters},
-  doi      = {https://doi.org/10.1016/j.cpc.2010.09.007},
+  doi      = {10.1016/j.cpc.2010.09.007},
   issn     = {0010-4655},
   number   = {12},
   pages    = {1929 - 1940},
@@ -130,7 +130,7 @@ url="https://doi.org/10.1007/128_2015_645"
   doi       = {10.1039/D0DT01404D},
   issue     = {24},
   pages     = {8086-8095},
-  url       = {http://dx.doi.org/10.1039/D0DT01404D},
+  url       = {10.1039/D0DT01404D},
   volume    = {49},
   abstract  = {Polynuclear transition metal complexes have continuously attracted interest owing to their peculiar electronic and magnetic properties which are influenced by the symmetry and connectivity of the metal centres. Understanding the full electronic picture in such cases often becomes difficult owing to the presence of multiple bridges between metal centres. We have investigated the electronic structure of a {Mn6} cage complex using computational and experimental approaches with the aim to understand the coupling between the manganese centres. The nature of the various coupling pathways has been determined using a novel methodology that involves perturbing the system while retaining the symmetry and analysing the effect on the coupling strength due to the perturbation. Furthermore{,} we have investigated the magnetic properties of this complex in higher oxidation states which reveals a switch in the nature of coupling from antiferromagnetic to ferromagnetic in addition to stabilisation of intermediate spin states.},
   journal   = {Dalton Trans.},
