### 2020-10-13 #2660 @dpsanders
👋 @Swetanshu: Thanks for your submission to JOSS!
Looking over the paper and repo, I have some preliminary comments:

- It's not clear why there are two standalone codes being submitted as one submission.

- Please expand the README in the repo with usage instructions and links to the documentation.

- It is not clear where the data from the DFT calculations comes from or what format it should be in. Please give a complete example.

- I do not see the source for the documentation -- this should be provided. It would be preferable for the docs to be available as web pages etc., rather than a monolithic PDF.

- Ideally, please include a simple usage example in the paper.

Thanks!
