+++ README.md
@@ -3,4 +3,4 @@ This package contains two standalone codes written in FORTRAN 2008 – ej_calc a
 The ej_calc code uses the spin densities obtained from the DFT calculations and determines the isotropic exchange coupling between paramagnetic centres. 
 The suscep code calculates the temperature dependence of magnetic susceptibility using the coupling constants.
 
-For a detailed description, please visit: https://github.com/WatsonGroupTCD/J2suscep/wiki
+For the documentation and detailed description of the codes, please visit: https://github.com/WatsonGroupTCD/J2suscep/wiki
