+++ README.md
@@ -1,6 +1,6 @@
 # J2suscep
 This package contains two standalone codes written in FORTRAN 2008 – ej_calc and suscep. 
-The ej_calc code uses the spin densities obtained from the DFT calculations and determines the isotropic exchange coupling between paramagnetic centres. 
+The ej_calc code uses the spin densities obtained from DFT calculations and determines the isotropic exchange coupling between paramagnetic centres. 
 The suscep code calculates the temperature dependence of magnetic susceptibility using the coupling constants.
 
-For the documentation and detailed description of the codes, please visit: https://github.com/WatsonGroupTCD/J2suscep/wiki
+For usage instructions and detailed documentation of the codes, please visit: https://github.com/WatsonGroupTCD/J2suscep/wiki
