+++ EJ_Calc/README.md
@@ -1,2 +1,3 @@
 # EJ_Calc
 Code for calculation of the exchange coupling constants
+For instructions, see https://github.com/WatsonGroupTCD/J2suscep/wiki 
