+++ Suscep/README.md
@@ -1,2 +1,3 @@
 # Suscep
 Code for calculation of magnetic susceptibility of molecular species using exchange coupling constants.
+For instructions, see https://github.com/WatsonGroupTCD/J2suscep/wiki 
