### 2020-11-05 #2660 @Swetanshu
Hi @dpsanders: Thank you for your comments and apologies for the late reply. I have tried to address your comments and the changes are described below:

•  It's not clear why there are two standalone codes being submitted as one submission.
The codes “EJ_calc” and “suscep” together provide a means to use the data obtained from density functional theory calculations for simulating the magnetic susceptibility. This requires the calculation of J-values (obtained from EJ_calc) followed by their use in the simulation of the susceptibility (obtained from the suscep code). Although, the two codes complement each other, the calculation of both J-values and susceptibility together is not always required. Additionally, the time required for the calculation of susceptibility can vary from seconds to hours. Keeping these things in mind, we decided to tackle the problem by developing two standalone codes instead of coupling them together.

•  Please expand the README in the repo with usage instructions and links to the documentation.
The readme file in the repository provides a brief description of the package. The detailed documentation has been provided in the github wiki pages. The README has been modified to indicate this.

•  It is not clear where the data from the DFT calculations comes from or what format it should be in. Please give a complete example.
An example has been included in the wiki pages in the example sections.

•  I do not see the source for the documentation -- this should be provided. It would be preferable for the docs to be available as web pages etc., rather than a monolithic PDF.
	The documentation in the pdf format has been left in the repo but all the information contained in the pdf files is also available in the github wiki pages.

•  Ideally, please include a simple usage example in the paper.
	Considering the length of the input files, the example covered in the paper is only briefly described. A detailed description of the same example is provided in the wiki pages. The input and output files for the same example are provided in the respective directories.

•  the software will need some kind of tests to make sure that it is working correctly.
	A few example input and output files are provided for this purpose. One can compare their output files with the provided ones to test the working of the code.

Let me know if you want me to modify anything else.
Thank you once again!
