### 2020-11-08 #2660 @mturiansky
@dpsanders I'd be happy to. This would be my first review for JOSS though, so please excuse my naivety.
