### 2020-11-11 #2660 @mfherbst
@dpsanders Sure, I'm happy to take it up. But same as @mturiansky it would be my first time reviewing for JOSS.
