### 2020-11-11 #2660 @dpsanders
Many thanks @mturiansky and @mfherbst!

Please head over to the review issue, where there's a checklist to orient your review.
The guidelines can be found in https://joss.readthedocs.io/en/latest/reviewer_guidelines.html

Please don't hesitate to ask if you have any questions. Thanks!
