### 2020-11-14 WatsonGroupTCD/J2suscep#1 @mturiansky
Installation
[_review for JOSS_](https://github.com/openjournals/joss-reviews/issues/2838)

Please provide installation instructions for your code. You should put this information in your wiki and clearly link to it in your README file. You should also clearly define the dependencies for this code.

From what I can glean, it needs a fortran compiler and LAPACK libraries? I would highly recommend including a single makefile (or some other compilation tool) at the base of your project, which compiles all the executables provided by the package.
