### 2020-11-14 WatsonGroupTCD/J2suscep#2 @mturiansky
Unit tests
[_review for JOSS_](https://github.com/openjournals/joss-reviews/issues/2838)

Please include unit tests to ensure the proper operation of your code. In the present form, I see no way to actually validate that the results returned from the code are correct. (We always need to look out for those pesky factors of 2pi :smile:)
