### 2020-11-14 WatsonGroupTCD/J2suscep#3 @mturiansky
formal vs. spin_dens
[_review for JOSS_](https://github.com/openjournals/joss-reviews/issues/2838)

Please describe the difference between the `EJ_Calc/spin_dens` directory and the `EJ_Calc/formal/` directory. Are they different codes? A large amount of the code is reproduced between the two. This is bad coding practice. One should identify the subroutines that are needed for both codes and move them to a separate file, which can then be linked to both executables. Reproducing code is very bad from a maintainability point of view.

The differences between the two codes needs to be explained better. I was under the impression that `ej_calc` and `suscep` were the two codes provided by this repository.
