### 2020-11-14 WatsonGroupTCD/J2suscep#4 @mturiansky
Contributing
[_review for JOSS_](https://github.com/openjournals/joss-reviews/issues/2838)

Please consider adding a `CONTRIBUTING.md` file or a section in the wiki on how people should go about contributing to the code. I see that there is a section for developers, but it is mainly an overview of the code itself, rather than instructions and best practices for contributing.
