### 2020-11-14 WatsonGroupTCD/J2suscep#5 @mturiansky
Example
[_review for JOSS_](https://github.com/openjournals/joss-reviews/issues/2838)

The example Mn complex is quite complex and has many degrees of freedom. It seems like a simpler system could be devised that would better illustrate how to use the code. I would imagine a two spin system with only one coupling constant should be feasible. (I understand that this doesn't show the major capability of the code of being able to sort out singular solutions and achieving some self-consistency, but I think it would be very useful for new users to see such an example.)
