### 2020-11-14 #2838 @mturiansky
@dpsanders a few quick questions: with regards to the "software paper" section, how strict is JOSS? Does there need to be an explicit section labeled "summary" or "state of the field". I feel that overall the paper covers these two topics, but not directly in their own section.

How about with writing quality? The paper gets its point across, but there are grammatical errors (run-on sentences, missing commas) that could be addressed.

I would say at this point that I have completed my review. I have opened several issues, and I would recommend major revisions. There is no installation instructions, so without major effort on my part, I cannot compile, run, or verify the output of the code. Also, there are no unit tests, which will likely require a not-insignificant effort to develop.
