### 2020-11-16 WatsonGroupTCD/J2suscep#1 @mfherbst
I agree with @mturiansky.

I tried to compile your code (replacing `ifort` by `gfortran` in the Makefiles as I don't have the Intel compilers) and did still not succeed as GNU make bailed out with an error. You should make sure that your compilation routines work out of the box for reasonable systems. Setting up continuous integration with one of the options available at github is from my experience very good for this purpose.
