### 2020-11-16 WatsonGroupTCD/J2suscep#2 @mfherbst
Again I agree with @mturiansky here. I would suggest using a simple example like the two-centre example suggested in #5 also for the purpose of unit testing as there the answer is probably easy to work out on pen and paper.
