### 2020-11-16 WatsonGroupTCD/J2suscep#6 @mfherbst
Structure of the repository
[review for JOSS](https://github.com/openjournals/joss-reviews/issues/2838)

I find the current structure of the code and the repository a little confusing. For example as has been pointed out in #3 the difference between `EJ_Calc/spin_dens` and `EJ_Calc/formal/` is not completely clear. Similarly multiple source code folders seem to contain example input or documentation which is furthermore appears to be duplicated with the wiki.

I would highly recommend to clearly separate source code, examples and documentation and in particular to not duplicate documentation as one instance is bound to become out of date. I would suggest to stick to one form of documentation and ideally one where the underlying tex / wiki or other markup code is part of the repository such that external contributors can easily improve it.
