### 2020-11-16 WatsonGroupTCD/J2suscep#3 @mfherbst
As far as I can capture some details are [given in the wiki](https://github.com/WatsonGroupTCD/J2suscep/wiki/1c_basic_details), but I would also recommend to rethink the structure of the repository.
