### 2020-11-16 WatsonGroupTCD/J2suscep#5 @mfherbst
I agree and suggest that a simpler example or a few simpler examples, which could also serve very well as unit test cases.
