### 2020-11-16 WatsonGroupTCD/J2suscep#7 @mfherbst
Simplification of workflow?
[review for JOSS](https://github.com/openjournals/joss-reviews/issues/2838)

I find it a bit strange that the project contains 3 different executables and it is not completely clear to me how the interoperate and which data that is output from one program should be fed to the other. To simply this I would suggest to restructure the workflow as follows:

Instead of shipping three separate executables, ship a single one that internally dispatches to functions doing the work for each stage of the calculation. In order to decide which steps should be executed one can use for example input arguments / commandline flags. If the data required for a particular step has not yet been computed it can be computed on the fly and passed from one stage to the other inside the executable without worrying the user to build a new input file manually for each stage of the calculation.

If the aforementioned is not possible for some reason than at the very least the authors should provide a set of script that execute the programs in the right order and parse / drop input / output files automatically to some extent to reduce on the required manual labour.
