### 2020-11-16 WatsonGroupTCD/J2suscep#8 @mfherbst
Continous Integration
[review for JOSS](https://github.com/openjournals/joss-reviews/issues/2838)

Along the points mentioned in #5 and #2 I would strongly suggest to setup some form of continuous integration to automatically compile and test the code. I have found this an indispensable tool for maintaining longer-running research codes myself and can only recommend it.
