### 2020-11-16 WatsonGroupTCD/J2suscep#9 @mfherbst
Example: Details about generation of data required for J2suscep
[review for JOSS](https://github.com/openjournals/joss-reviews/issues/2838)

The required data for setting up J2suscep, namely results from population analysis / DFT energies etc. are in itself not always easy to obtain (especially for broken symmetry states). I think it would make the documentation more self-contained if the authors could link to manual sections of appropriate quantum-chemistry codes to do the calculations to detail the users how to produce such data. Even better the authors could provide a simple walked-through tutorial explaining exactly for one quantum chemistry code what calculations and input options have to be chosen to generate the data required for J2suscep.
