### 2020-11-16 WatsonGroupTCD/J2suscep#10 @mfherbst
Difference / relation of J2suscep to other codes
[review for JOSS](https://github.com/openjournals/joss-reviews/issues/2838)

In reading the JOSS paper it did not come across to me what the relationship / difference of J2suscep to other codes such as PHI and MAGPACK is. Especially PHI seems to be a pretty sophisticated code from what I capture. I would recommend that the authors improve on that section of the paper and strengthen the need for developing J2suscep.
