### 2020-11-16 #2838 @mfherbst
@dpsanders I am now also done with my review and would recommend major revisions. I opened up various issues with respect to my points of critique or added some comments to the earlier ones of @mturiansky.

Similar to @mturiansky I wonder a little about the Software paper section. I found the summary and statement of need addressed
to a satisfactory point, albeit not in particularly labelled sections. In contrast I am not so sure about the description of the state of the art. While I am not an expert in this particular subfield of quantum chemistry, it seems to me the authors solve a common enough problem, such that it could be worth expanding on the relationship of their code compared to other solutions in the field. Since JOSS is more centred around the code rather than the scientific methods, I am not sure how strict JOSS is in this respect, though.

Another aspect is the quality of the code itself. From the snippets I read the code makes use of some features that might make it difficult to maintain in the long run (e.g. global variables). How much is this a concern for JOSS? Is it about the functionality of the present stage of the code or is it also about the long-term maintainability and the value a project adds to the open-source community?

Despite clearly involving some effort I think the code at its present stage could benefit a lot from taking the time to come up with a better structure, such that e.g. code duplication or global variables are avoided. This would not only make it more accessible for others, but also easier to maintain and test.
