### 2020-12-04 WatsonGroupTCD/J2suscep#2 @dpsanders
JOSS requires that there are tests, but they strictly do not need to be automated.
However, I agree that setting up continuous integration is key for modern software projects.
