### 2020-12-04 WatsonGroupTCD/J2suscep#11 @dpsanders
Move docs from wiki into main repo
The wiki will not be archived by Zenodo, so you will need to move (or at least copy) the files from there into the main repo, e.g. in a folder called `docs`.

You can clone the wiki using the link in the bottom right corner at https://github.com/WatsonGroupTCD/J2suscep/wiki, which will give you the markdown files.

Part of [review for JOSS](https://github.com/openjournals/joss-reviews/issues/2838)
