### 2020-12-04 WatsonGroupTCD/J2suscep#4 @dpsanders
This is a requirement for JOSS.
