### 2020-12-04 WatsonGroupTCD/J2suscep#9 @dpsanders
+1 for a tutorial.

I had already requested a worked example in the pre-review.
