### 2020-12-04 #2838 @dpsanders
👋 Hi @mturiansky and @mfherbst: Many thanks for your detailed and perceptive comments and issues on the repo. My apologies for the long delay in getting to this.

I agree with all your points (modulo a couple of clarifying remarks in the issues). @swetanshu: I would ask you to please address the points raised in the issues on the repo, which will make for a more useful and reusable piece of software. I will respond to particular points below.
