### 2020-12-04 WatsonGroupTCD/J2suscep#12 @dpsanders
Authorship clarification for JOSS paper
It seems that only @Swetanshu has contributed code to the repo.
Could you please clarify the contributions of the other two authors to the JOSS paper?

Part of [review for JOSS](https://github.com/openjournals/joss-reviews/issues/2838)
