### 2020-12-04 #2838 @dpsanders
@mturiansky:

> with regards to the "software paper" section, how strict is JOSS? Does there need to be an explicit section labeled "summary" or "state of the field". I feel that overall the paper covers these two topics, but not directly in their own section.

HOSS has no explicit requirements about this, but I agree that that would improve the structure of the paper.

> How about with writing quality? The paper gets its point across, but there are grammatical errors (run-on sentences, missing commas) that could be addressed.

The text should be readable; if it isn't, that is unacceptable. I am happy to help with minor grammar and punctuation issues once the paper is ready.

> There is no installation instructions, so without major effort on my part, I cannot compile, run, or verify the output of the code. Also, there are no unit tests, which will likely require a not-insignificant effort to develop.

Installation instructions are a requirement of JOSS, since without them the software is, as you say, unusable. It is a requirement for their to be tests; the tests are not *required* to be automated, but I agree that this is a standard feature in modern software.

Thanks again for your comments!
