### 2020-12-04 #2838 @dpsanders
@mfherbst:

> I am not so sure about the description of the state of the art. While I am not an expert in this particular subfield of quantum chemistry, it seems to me the authors solve a common enough problem, such that it could be worth expanding on the relationship of their code compared to other solutions in the field. Since JOSS is more centred around the code rather than the scientific methods, I am not sure how strict JOSS is in this respect, though.

It is a requirement to describe the relation of the software with other common packages in the same area. Implicitly that requires specifying how this particular piece of software is different from the others; this is related to the Statement of need, which needs to say why *this* particular piece of software is required.

In particular, https://joss.readthedocs.io/en/latest/submitting.html states that "[the] software should be a significant contribution to the available open source software that either enables some new research challenges to be addressed or makes addressing research challenges significantly better (e.g., faster, easier, simpler)." It should thus be explicitly addressed in the paper in what way this is accomplished.

> Another aspect is the quality of the code itself. From the snippets I read the code makes use of some features that might make it difficult to maintain in the long run. Despite clearly involving some effort I think the code at its present stage could benefit a lot from taking the time to come up with a better structure, such that e.g. code duplication or global variables are avoided. This would not only make it more accessible for others, but also easier to maintain and test.

> How much is this a concern for JOSS? Is it about the functionality of the present stage of the code or is it also about the long-term maintainability and the value a project adds to the open-source community?

It is a requirement (see the above-cited page) that "JOSS requires that software should be feature-complete... and designed for maintainable extension... single-function packages are not acceptable."

So I agree with you that the software should be written in a way that enables other people to understand its structure and maintain it. If you think the current structure of the code does not allow that then it needs to be modified for publication in JOSS.

Thanks again for your comments!
