### 2020-12-10 #2838 @Swetanshu
Thank you @mturiansky and @mfherbst for the quick review, and @dpsanders for your response!

We have been working on updating the code in the past few days. I was not sure whether I should make any commits before @dpsanders gets back so I did not do that. I’ll start committing the updates that have been made. To summarise, the code has been restructured, and examples and documentation have been updated. We are currently working on implementing automated tests. I’ll get back to you as soon as that is done.
