+++ EJ_Calc/formal/init.f90
@@ -1,258 +1,0 @@
-Module init
-
- Use Iso_Fortran_Env
- Implicit None
- Integer, parameter :: int_kind = INT32
- Integer, parameter :: real_kind = REAL64 !Double precision
- Integer, parameter :: comp_kind = REAL64        
- Integer(kind = int_kind) :: i, j, num_mag_cent, no_of_j_val, tot_interac
- !i and j are for running loops
- !num_mag_cent stores the number of metal centres in each configuration
- !no_of_j_val stores the total number of Jvalues defined for the system
- !tot_interac stores the number of possible interactions between spins i.e. the
- !total number of S1S2 terms.
- Integer(kind = int_kind) :: ios, num_spin_dens_set, try, try1, poss_comb
- !num_spin_dens_set stores the number of sets of spin density available
- !which is the total numberof lines in the spin density file
- !poss_comb determined the possible combination of equations i.e. the
- !(n-1)Cj term (read line 10 from top)
- !try and try1 help in keeping the program running just in case there is an
- !equation set is encountered that cannot be solved.
- Integer(kind = int_kind), Dimension(:), Allocatable :: hamil1, hamil2, jpos
- !jpos keeps track of how many interactions are being included under a single J
- !value
- !hamil1 and hamil2 together keep track of which centres are interacting
- Integer(kind = int_kind) :: main_cntr
- !main_cntr will keep track of total non singular equations
- Real(kind = real_kind), Dimension(:), Allocatable :: energy, tot_avg
- !energy stores the energy for each electronic configuration
- !tot_avg stores the average of all the J-value sets which are valid and which
- !meet the standard deviation criteria.
- Real(kind = real_kind), Dimension(:), Allocatable :: stddev, std_per
- !stddev and std_per store the standard deviation and percentage standard
- !deviation for the different J-values
- Real(kind = real_kind), Dimension(:,:), Allocatable :: spin, hamil, jval
- !spin stores the spin density value associated with each metal centre as
- !given in the spin density file.
- !hamil stores the 2S1S2 type weight terms for each interaction. It helps in
- !determining which J value is associated with which S1S2 term.
- !jval stores the total (2S1S2 type )coefficient associated with each J value
- !Each column represents a different J value for jval
- Real(kind = real_kind), Dimension(:,:,:), Allocatable :: jval_trkr
- !jval_trkr stores all the computed sets of J values
- Real(kind = real_kind) :: start, finish
- !start and finish help determine the time taken during the program execution
- Character (Len = 100) :: FileName1, FileName2, read_line, Outfile
- Logical :: EndOfFile, singular, normal, identical, stdv
- !Call cpu_time(start)
- !!Call init()
-
-Contains
-!!!Initialization Subroutines!!!
-         Subroutine Init1()
-           !!!!! Declaration and initailisation of local and global variables
-           !!!!!!
-                Logical :: check, check1
-                Integer(kind = int_kind) :: pos
-                !pos helps in defining the different J values.
-                Real(kind = real_kind) :: val
-                check = .False.
-                check1 = .False.
-                168 Format(i2)
-                189 Format(F4.1)
-                193 Format(F6.3)
-                245 Format(F10.7)
-                249 Format(F10.6)
-           !!!!! Declaration and initailisation ends !!!!!
-
-           !!!!! Determining number of magnetic centres (i.e. metal atoms) !!!!!
-           !!!!! Initialisation of spin for each metal centre !!!!!
-                Read(11, '(a)', iostat = ios) read_line
-                If(trim(read_line) == 'magnetic centres') Then
-                        Read(11, 168) num_mag_cent
-                Else
-                        Write(*,*) ' Erroneous number of magnetic centres in the input file'
[…]
