### 2020-12-12 ea12233
Suscep input file matches ej_calc. Suscep examples updated to match those of ej_calc.
