+++ src/suscep/suscep.f90
@@ -28,8 +28,8 @@ Program suscep
  ! posval description given in the subroutine FormBasis
  Integer(kind = int_kind), Dimension(:), Allocatable :: pseudospin
  ! pseudospin is simply a pointer pointing to the position of each metal centre
- Integer(kind = int_kind), Dimension(:,:), Allocatable :: jmat
- ! jmat is the J-value matrix storing information about which J-values 
+ Integer(kind = int_kind), Dimension(:,:), Allocatable :: jmatx, jmaty
+ ! jmatx and jmaty are the J-value matrix storing information about which J-values 
  !are associated with different spin pairs
  Real (kind = real_kind), Dimension(:,:), Allocatable :: hamil 
  ! hamil is the spin hamiltonian matrix
@@ -71,7 +71,7 @@ Program suscep
  Call FormBasis()
  Call HamilForm(totalspin)
  Deallocate (spin, basis, pseudospin, hamil)
- Deallocate (spinmat, posval, jval, jmat)
+ Deallocate (spinmat, posval, jval, jmatx, jmaty)
  Close(11)
  Close(12)
 
@@ -103,7 +103,7 @@ Program suscep
                 Read(11, 168) dmension
         Else 
                 Write(*,*) ' Erroneous number of magnetic centres in the input file' 
-                Call Abort
+                STOP 
         End If
         Write(12,*) "Number of Magnetic centers:", dmension
         Allocate(spin(dmension))
@@ -125,7 +125,7 @@ Program suscep
         Do i = 1, dmension
                 If(spin(i) == 0) Then
                         Write(*,*) "Erroneous spin values"
-                        Call Abort
+                        STOP 
                 End If
         End Do
    !!!!! Reading of spins ends !!!!!
@@ -136,7 +136,7 @@ Program suscep
                 Read(11, 168) no_of_j_val 
         Else
                 Write(*,*) ' Erroneous No. of J values'
-                Call Abort
+                STOP 
         End If
         Allocate(jval(no_of_j_val))
         Read(11, '(a)', iostat = ios) read_line
@@ -149,7 +149,7 @@ Program suscep
                 End Do
         Else
                 Write(*,*) ' Erroneous J values'
-                Call Abort
+                STOP 
         End If
         Write(12,*)
    !!!!! J value determination ends !!!!!
@@ -161,7 +161,7 @@ Program suscep
                 Write(12,*) 'g value:', g
         Else
                 Write(12,*) 'Erroneous g value'
-                Call Abort
+                STOP
         End If
    !!!!! G value determination ends!!!!!
 
@@ -171,68 +171,56 @@ Program suscep
                 Read(11, '(a)', iostat = ios) read_line
         Else
                 Write(*,*) 'No Hamiltonian'
-                Call Abort
+                STOP 
         End If
 
    !!!!! Determining the number of variables in the Hamiltonian !!!!!
-        Do i = 1, len_trim(read_line)
-                If(read_line(i:i) == '(') Then
-                        check = .Not.check
-                End If
-                If (check .eqv. .True.) Then
-                        if (read_line(i:i) == '+') Then
-                                i1 = i1 + 1
-                        End If
-                End If
-                If (read_line(i:i) == ')') Then
-                        check = .Not.check
+        i1 = 0
+        Do While (trim(read_line) .ne. 'Field Strength')
+                If(trim(read_line) == '****' .OR. trim(read_line) == 'Field Strength') Then
                         If (i1 > j1) Then
                                 j1 = i1
-                                i1 = 0
                         End If
+                        i1 = 0
+                Else
+                        i1 = i1 + 1 
                 End If
[…]
