+++ src/suscep/suscep.f90
@@ -99,7 +99,7 @@ Program suscep
    !!!!! Determining number of magnetic centres (i.e. metal atoms) !!!!!
    !!!!! Initialisation of spin for each metal centre !!!!!
         Read(11, '(a)', iostat = ios) read_line
-        If(trim(read_line) == 'magnetic centres') Then
+        If(trim(read_line) .eq. 'magnetic centres') Then
                 Read(11, 168) dmension
         Else 
                 Write(*,*) ' Erroneous number of magnetic centres in the input file' 
@@ -115,7 +115,7 @@ Program suscep
 
    !!!!! Reading in the actual values of spin for each centre !!!!!
         Read(11, '(a)', iostat = ios) read_line
-        If(trim(read_line) == 'spin') Then
+        If(trim(read_line) .eq. 'spin') Then
                 Do i = 1, dmension
                         Read(11, 189) spin(i)
                         pseudospin(i) = i
@@ -123,7 +123,7 @@ Program suscep
                 End Do
         End If
         Do i = 1, dmension
-                If(spin(i) == 0) Then
+                If(spin(i) .eq. 0) Then
                         Write(*,*) "Erroneous spin values"
                         STOP 
                 End If
@@ -132,7 +132,7 @@ Program suscep
 
    !!!!! Determining the number of J values, reading them and printing them!!!!!
         Read(11, '(a)', iostat = ios) read_line
-        If(trim(read_line) == 'No. of J values') then
+        If(trim(read_line) .eq. 'No. of J values') then
                 Read(11, 168) no_of_j_val 
         Else
                 Write(*,*) ' Erroneous No. of J values'
@@ -140,7 +140,7 @@ Program suscep
         End If
         Allocate(jval(no_of_j_val))
         Read(11, '(a)', iostat = ios) read_line
-        If(trim(read_line) == 'J values') then
+        If(trim(read_line) .eq. 'J values') then
                 Write(12,'(a)', advance = 'no') ' The J values(cm-1) given are:'
                 Do i = 1, no_of_j_val
                         Read(11, 245) jval(i)
@@ -156,7 +156,7 @@ Program suscep
 
    !!!!! Determining the g value!!!!!
         Read(11, '(a)', iostat = ios) read_line
-        If (trim(read_line) == 'g value') Then
+        If (trim(read_line) .eq. 'g value') Then
                 Read(11, 193) g
                 Write(12,*) 'g value:', g
         Else
@@ -167,7 +167,7 @@ Program suscep
 
    !!!!! Reading in and printing the Spin Hamiltonian !!!!!
         Read(11, '(a)', iostat = ios) read_line
-        If(trim(read_line) == 'Hamiltonian') Then
+        If(trim(read_line) .eq. 'Hamiltonian') Then
                 Read(11, '(a)', iostat = ios) read_line
         Else
                 Write(*,*) 'No Hamiltonian'
@@ -177,8 +177,8 @@ Program suscep
    !!!!! Determining the number of variables in the Hamiltonian !!!!!
         i1 = 0
         Do While (trim(read_line) .ne. 'Field Strength')
-                If(trim(read_line) == '****' .OR. trim(read_line) == 'Field Strength') Then
-                        If (i1 > j1) Then
+                If((trim(read_line) .eq. '****') .OR. (trim(read_line) .eq. 'Field Strength')) Then
+                        If (i1 .gt. j1) Then
                                 j1 = i1
                         End If
                         i1 = 0
@@ -187,7 +187,7 @@ Program suscep
                 End If
                 Read(11, '(a)', iostat = ios) read_line
         End Do
-        If (i1 > j1) Then
+        If (i1 .gt. j1) Then
                 j1 = i1
[…]
