+++ examples/test/Makefile
@@ -1,18 +1,0 @@
-
-.PHONY: test test_ej_calc_form test_ej_calc_spin  test_suscep        
-
-test:  test_ej_calc_form test_ej_calc_spin test_suscep 
-
-test_ej_calc_form:         
-	./test.sh ej_calc_form
-
-test_ej_calc_spin:         
-	./test.sh ej_calc_spin 
-
-test_suscep:         
-	./test.sh suscep 
-
-clean:
-	rm -fr ej_calc_form ej_calc_spin suscep 
-
-  
