### 2021-01-25 fcd452a
minor update: modul_lib; removed a comment
