### 2021-02-05 82c75b2
Streamline input
