+++ JOSS_paper/paper.md
@@ -27,12 +27,12 @@ bibliography: paper.bib
 # `Statement of need`
 
 The field of molecular magnetism has attracted significant interest owing 
-to the rising need to miniaturize magnets. Advancement in this field 
-requires a deep understanding of the electronic structure, and computational 
+to the rising need to miniaturize magnets. Advancements in this field 
+require a deep understanding of the electronic structure, and computational 
 approaches, in this respect, have played a key role in pushing the limits 
 further [@Neese]. A deep understanding of the magnetic properties not 
-only helps in making smaller magnets but has also been instrumental in 
-speculating the structure of systems that are difficult to crystallise or 
+only helps in making magnets smaller, but has also been instrumental in 
+speculating the structure of systems that are difficult to crystallise, or 
 of intermediate states during a catalytic process [@Krewald].
 
 In molecular complexes, the magnetic interaction between paramagnetic 
@@ -66,7 +66,7 @@ obtained from any approach can be used for the calculation of J-values. The code
 calculates all possible solutions based on the Hamiltonian and the states modelled, 
 removes any singular solutions and calculates an average set of coupling constants 
 and the standard deviations. Additionally, it also calculates the energy of the 
-different spin states relative to each other using the coupling constants for 
+different spin states relative to each other using the coupling constants, for 
 comparison to the original DFT data.
 
 The suscep code calculates the temperature dependence of magnetic susceptibility 
@@ -85,7 +85,7 @@ for the calculation of the 2 J-values results in 60 possible solutions and using
 the final values were determined to be -1.28 (cis-coupling) and -3.48 cm^-1^ (trans-coupling) 
 respectively [@Mn6]. The use of the suscep code calculates the temperature dependence of 
 susceptibility which is shown in Figure 1 (b). This package has been used in a similar manner 
-for a {$Mn_8$} complex [@Mn8].
+for probing the magnetic properties of a {$Mn_8$} complex [@Mn8].
 
 ![(a) Structure of the {$Mn_6$} complex and (b) the temperature dependence of magnetic 
 susceptibility obtained using the suscep code. Colour scheme: Mn (dark blue), P (pink), Cl (green), 
@@ -96,6 +96,7 @@ C (black), N (blue) and O (red). Hydrogen atoms have been removed for clarity.](
 
 The authors are grateful to the Irish research council (GOIPG/2015/2952), Science foundation 
 Ireland (12/IA/1414 and 13/IA/1896) and European research council (CoG 2014–647719) for the funding.
+The authors also  acknowledge ICHEC for the computational resources.  
 
 
 # References
