+++ JOSS_paper/paper.bib
@@ -21,6 +21,35 @@ address = {USA}
   year     = {2013},
 }
 
+@inbook{Bsym,
+author = {Noodleman, Louis and Case, David A.},
+publisher = {American Cancer Society},
+isbn = {9780470862100},
+title = {Broken Symmetry States of Iron—Sulfur Clusters},
+booktitle = {Encyclopedia of Inorganic Chemistry},
+chapter = {},
+pages = {},
+doi = {https://doi.org/10.1002/0470862106.ia618},
+url = {https://onlinelibrary.wiley.com/doi/abs/10.1002/0470862106.ia618},
+eprint = {https://onlinelibrary.wiley.com/doi/pdf/10.1002/0470862106.ia618},
+year = {2009},
+keywords = {iron–sulfur proteins, ferredoxins, nitrogenase, DFT, spin coupling, spin-dependent delocalization, redox potentials, broken symmetry, electrostatics, ENDOR, hyperfine, Heisenberg coupling parameter},
+abstract = {Abstract After presenting a general introduction to the structures and physical properties of dinuclear and polynuclear iron sulfur clusters, we discuss the relationship between the broken symmetry model within density functional theory (DFT), and the energies and properties of related pure spin states. The methodology for calculating and the physical significance of Heisenberg coupling parameters (J) and spin-dependent delocalization parameters (B) are developed. The interaction energy of the active site quantum cluster with the protein and solvent environment is treated with a single step procedure based on the classical Poisson–Boltzmann (PB equation). Redox potentials of various high-potential 4Fe4S and ferredoxin proteins are calculated with DFT/PB methods. The electronic and spin-coupling structure of oxidized high-potential (HIPIP) proteins is examined, and it is shown that two distinct electronic structures are possible. These contrasting electronic structures may have implications for the structural stability of oxidized HIPIP. For the Fe protein of nitrogenase, the redox potential and spin state of the super-reduced form are calculated, and it is argued that this all-ferrous state is probably not physiologically accessible. The redox potential and electronic structure of the FeMo cofactor of the MoFe protein of nitrogenase are examined for the “resting” state. A very small spin density at the central atom is very feasible. The central atom could be XC, N, O, but C, N are most probable, despite the absence of an observable ENDOR signal for the central atom. The spin-density for a central atom should be much higher for the 2e-reduced FeMo cofactor when a ligand is bound, and we have tested this for bound allylic alcohol. Three different ligand-cluster binding structures are characterized by geometry, energy, and spin distribution.}
+}
+
+@article{Bsym1,
+author = {Ruiz, Eliseo and Cano, Joan and Alvarez, Santiago and Alemany, Pere},
+title = {Broken symmetry approach to calculation of exchange coupling constants for homobinuclear and heterobinuclear transition metal complexes},
+doi = {10.1002/(SICI)1096-987X(199910)20:13<1391::AID-JCC6>3.0.CO;2-J}
+eprint = {https://onlinelibrary.wiley.com/doi/pdf/10.1002/%28SICI%291096-987X%28199910%2920%3A13%3C1391%3A%3AAID-JCC6%3E3.0.CO%3B2-J},
+journal = {Journal of Computational Chemistry},
+volume = {20},
+number = {13},
+pages = {1391-1400},
+url = {https://onlinelibrary.wiley.com/doi/abs/10.1002/%28SICI%291096-987X%28199910%2920%3A13%3C1391%3A%3AAID-JCC6%3E3.0.CO%3B2-J},
+year = {1999}
+}
+
 @Article{Cremades1,
[…]
