+++ JOSS_paper/paper.bib
@@ -41,7 +41,7 @@ abstract = {Abstract After presenting a general introduction to the structures a
 author = {Ruiz, Eliseo and Cano, Joan and Alvarez, Santiago and Alemany, Pere},
 title = {Broken symmetry approach to calculation of exchange coupling constants for homobinuclear and heterobinuclear transition metal complexes},
 doi = {10.1002/(SICI)1096-987X(199910)20:13<1391::AID-JCC6>3.0.CO;2-J}
-eprint = {https://onlinelibrary.wiley.com/doi/pdf/10.1002/%28SICI%291096-987X%28199910%2920%3A13%3C1391%3A%3AAID-JCC6%3E3.0.CO%3B2-J},
+eprint = {https://doi.org/10.1002/(SICI)1096-987X(199910)20:13<1391::AID-JCC6>3.0.CO;2-J},
 journal = {Journal of Computational Chemistry},
 volume = {20},
 number = {13},
