+++ JOSS_paper/paper.bib
@@ -6,6 +6,7 @@ isbn = {0898714478},
 publisher = {Society for Industrial and Applied Mathematics},
 address = {USA}
 }
+
 @Article{phi,
   author   = {Chilton, Nicholas F. and Anderson, Russell P. and Turner, Lincoln D. and Soncini, Alessandro and Murray, Keith S.},
   title    = {PHI: A powerful new program for the analysis of anisotropic monomeric and exchange-coupled polynuclear d- and f-block complexes},
@@ -38,16 +39,16 @@ abstract = {Abstract After presenting a general introduction to the structures a
 }
 
 @article{Bsym1,
-author = {Ruiz, Eliseo and Cano, Joan and Alvarez, Santiago and Alemany, Pere},
-title = {Broken symmetry approach to calculation of exchange coupling constants for homobinuclear and heterobinuclear transition metal complexes},
-doi = {10.1002/(SICI)1096-987X(199910)20:13<1391::AID-JCC6>3.0.CO;2-J}
-eprint = {https://doi.org/10.1002/(SICI)1096-987X(199910)20:13<1391::AID-JCC6>3.0.CO;2-J},
-journal = {Journal of Computational Chemistry},
-volume = {20},
-number = {13},
-pages = {1391-1400},
-url = {https://onlinelibrary.wiley.com/doi/abs/10.1002/%28SICI%291096-987X%28199910%2920%3A13%3C1391%3A%3AAID-JCC6%3E3.0.CO%3B2-J},
-year = {1999}
+  author = {Ruiz, Eliseo and Cano, Joan and Alvarez, Santiago and Alemany, Pere},
+  title = {Broken symmetry approach to calculation of exchange coupling constants for homobinuclear and heterobinuclear transition metal complexes},
+  doi = {10.1002/(SICI)1096-987X(199910)20:13<1391::AID-JCC6>3.0.CO;2-J},
+  eprint = {https://doi.org/10.1002/(SICI)1096-987X(199910)20:13<1391::AID-JCC6>3.0.CO;2-J},
+  journal = {Journal of Computational Chemistry},
+  volume = {20},
+  number = {13},
+  pages = {1391-1400},
+  url = {https://onlinelibrary.wiley.com/doi/abs/10.1002/%28SICI%291096-987X%28199910%2920%3A13%3C1391%3A%3AAID-JCC6%3E3.0.CO%3B2-J},
+  year = {1999}
 }
 
 @Article{Cremades1,
