+++ JOSS_paper/paper.bib
@@ -168,6 +168,20 @@ url="https://doi.org/10.1007/128_2015_645"
   year      = {2020},
 }
 
+@article{spin_dens,
+author = {Paul, Satadal and Misra, Anirban},
+title = {Interpretation and Quantification of Magnetic Interaction through Spin Topology},
+journal = {Journal of Chemical Theory and Computation},
+volume = {8},
+number = {3},
+pages = {843-853},
+year = {2012},
+doi = {10.1021/ct2006506},
+note ={PMID: 26593346},
+URL = {https://doi.org/10.1021/ct2006506},
+eprint = {https://doi.org/10.1021/ct2006506}
+}
+
 @Article{Vignesh,
   author   = {Vignesh, Kuduva R. and Langley, Stuart K. and Murray, Keith S. and Rajaraman, Gopalan},
   title    = {What Controls the Magnetic Exchange Interaction in Mixed- and Homo-Valent $Mn_7$ Disc-Like Clusters? A Theoretical Perspective},
