### 2021-02-12 WatsonGroupTCD/J2suscep#1 @Swetanshu
Installation instructions have been provided in the wiki now. A single makefile is required to compile and generate all the executables. The makefile takes in an include file that contains all the compiler specific options. An arch directory containing model include files for GNU and Intel compilers has also been included. Additionally, the blas options are also included in the model include files for different blas libraries.
