### 2021-02-12 WatsonGroupTCD/J2suscep#2 @Swetanshu
Automated tests have been implemented. These tests, use the examples provided in the package and verify that the compiled codes generate results similar to those provided with the examples. As some of the examples take significantly longer to run compared to others, the tests have been implemented in a way that one can selectively run shorter and longer tests (details described in the installation instructions).
