### 2021-02-12 WatsonGroupTCD/J2suscep#3 @Swetanshu
The ej_calc code has been restructured to have a single source code. The compilation of the code produces two executables ej_calc_spin and ej_calc_form. The usage of both executables has been descibed in the wiki.
