### 2021-02-12 WatsonGroupTCD/J2suscep#5 @Swetanshu
Simpler examples have been provided as part of the package. These examples have also been discussed in the wiki and are being used as test cases.
