### 2021-02-12 WatsonGroupTCD/J2suscep#6 @Swetanshu
The entire repository has been restructured. Separate source code, examples and documentation directories have been created. The duplicate documentation and duplicate code sections have been removed.
