### 2021-02-12 WatsonGroupTCD/J2suscep#7 @Swetanshu
Considering that the execution of suscep requires additional information from the user, and that the calculation of both J-values and susceptibility is not necessarily required every time, we decided not to create a single executable.  Clear instructions to execute each executable have been provided in the wiki. We have also included a basic script (in the bin directory) for executing both the ej_calc executables.
