### 2021-02-12 WatsonGroupTCD/J2suscep#8 @Swetanshu
We have implemented automated tests for validating the code.
