### 2021-02-12 WatsonGroupTCD/J2suscep#9 @Swetanshu
A walked-through tutorial for modelling the broken symmetry states using the Gaussian software has been provided in the documentation (see appendix).
