### 2021-02-12 WatsonGroupTCD/J2suscep#10 @Swetanshu
We already mention in the paper that the application covered by ej_calc is not implemented in any other code. We have included the suscep code, which shares some similarity with PHI and MAGPACK, to make the package self-contained.
