+++ documentation/1a_introduction.md
@@ -1,0 +1,19 @@
+<html><head></head><body>
+<h1>Introduction</h1>
+<p>Briefly, this code calculates coupling constants using the following Hamiltonian:
+<font face="Arial, Helvetica, sans-serif" size="-2">[<a href="refs#Heisenberg" class="showTip Heisenberg">Heisenberg</a>], [<a href="refs#Dirac" 
+class="showTip Dirac">Dirac</a>], [<a href="refs#Vanvleck" class="showTip Vanvleck">Vanvleck</a>]</font>  </p> 
+<br>
+
+<font face="Courier New, Courier, monospace" size="-2">&nbsp;&nbsp;&nbsp;&nbsp;</font>H&#770; = -2 &sum;<sub>i>j</sub> J<sub>ij</sub>s<sub>i</sub>s<sub>j</sub>	
+<p>where the terms have the following meanings:</p>
+<table border="0" cellpadding="2">
+<tr valign="top"><td><p>J<sub>ij</sub></p></td><td>&nbsp;&nbsp;</td>
+<td><p>The exchange coupling constant</p></td></tr>
+<tr valign="top"><td><p>s<sub>x</sub></p></td><td>&nbsp;&nbsp;</td>
+<td><p>The spin operator.</p></td></tr>
+</table>
+<p> Any definition of spin operators (formal spins or spin projections or spin densities or any other) can be used. Sometimes researchers exclude the pre-factor 2 in the above formula so this needs to be kept in mind while using this code.</p>
+
+    
+</body></html>
