+++ documentation/2b_theory.md
@@ -1,0 +1,86 @@
+<html><head></head><body>
+<h1>Theory</h1>
+
+<p></p>
+
+<p><b>Matrix Elements:</b> The matrix elements are constructed using an uncoupled basis comprising of spin elements (the magnetic spin quantum number, M<sub>s</sub>, value) 
+for each magnetic centre. If a magnetic centre has N unpaired electron, then the magnetic spin quantum number values for that particular magnetic centre range from N/2 
+to –N/2; the consecutive values differing from each other by unity (total (N+1) values).</p>
+
+<p><b>Hamiltonian:</b> Only the two major contributors – the isotropic exchange term and the Zeeman term – to the Hamiltonian have been taken into consideration.</p>
+<p>Isotropic Exchange term:<font face="Arial, Helvetica, sans-serif" size="-2">[<a href="refs#Boca1" class="showTip Boca1">Boca1</a>]</font>  </p>
+
+<p align="center"><img src="img/two.jpg" alt="eq. 2"><br /></p>
+<p>where <i>J<sub>AB</sub></i> is the coupling constant between the magnetic centres A and B, <i>I</i> and <i>J</i> represent the particular M<sub>s</sub> terms in the basis elements 
+that <i>S&#770<sub>A</sub>.S&#770<sub>B</sub></i> operate on and</p>
+
+<p align="center"><img src="img/three.jpg" alt="eq. 3"><br /></p>
+<p>where <i>S&#770<sub>A,+</sub></i> and <i>S&#770<sub>B,-</sub></i> are increment and decrement operators that are defined as follows: </p>
+
+<p align="center"><img src="img/four_six.jpg" alt="eq. 4-6"><br /></p>
+<br>
+
+<p>Zeeman term:<font face="Arial, Helvetica, sans-serif" size="-2">[<a href="refs#Boca1" class="showTip Boca1">Boca1</a>]</font> </p>
+<p align="center"><img src="img/seven.JPG" alt="eq. 7"><br /></p>
+<p>Here <i>B&#8407</i> represents the applied magnetic field, g<sub>A</sub> is the gyromagnetic tensor and μ<sub>B</sub> is the unit Bohr Magneton.</p>
+
+<p align="center"><img src="img/eight.jpg" alt="eq. 8"><br /></p>
+<p>where δ<sub>AB</sub> is the kroenecker delta operator which is equal to 1 if A=B. Otherwise its value is 0. <i>B<sub>i</sub></i> and g<sub>Ai</sub> represent the value of 
+the magnetic field and the gyromagnetic tensor along a given direction. The effective Hamiltonian is thus given as </p>
+
+<p align="center"><img src="img/nine.JPG" alt="eq. 9"><br /></p>
+<p>The Hamiltonian matrix can thus be represented as </p>
+
+<p align="center"><img src="img/ten.JPG" alt="eq. 10"><br /></p>
+<p>where <i>A<sub>I</sub></i> and <i>A<sub>J</sub></i> represent the elements of the basis and N is the total number of elements in the basis. </p>
+
+<p>The eigenvalues ε<sub>i</sub>  and hence the energy levels of the various magnetic states are obtained by the diagonalisation of the Hamiltonian matrix. 
+Once the eigenvalues are obtained, it can be used to plot the magnetic susceptibility with respect to temperature. To calculate the magnetic susceptibility, 
+the Van Vleck equation is used which is:<font face="Arial, Helvetica, sans-serif" size="-2">[<a href="refs#Kahn" class="showTip Kahn">Kahn</a>]</font> </p>
[…]
