+++ documentation/2g_limitations.md
@@ -1,0 +1,17 @@
+<html><head></head><body>
+<h1>Limitations</h1>
+
+
+<p>The determination of the susceptibility for very large systems is limited by the available memory (RAM). This is because the size of the matrix 
+containing M<sub>s</sub> elements increases rapidly with the increase in the number of unpaired electrons and paramagnetic centres. The total 
+number of elements in the matrix containing the M<sub>s</sub> elements can be given as [(no. of unpaired electrons on centre 1) * (no. of unpaired 
+electrons on centre 2) * (no. of unpaired electrons on centre 3) * ……...]<sup>2</sup>. For example, for a {Mn<sup>III</sup><sub>4</sub>} system, 
+the matrix contains a total of (5*5*5*5)<sup>2</sup> =) 390,625 elements, for a {Mn<sup>III</sup><sub>5</sub>} system, the matrix contains a total 
+of (5<sup>5</sup>*5<sup>5</sup> =) 9,765,625 elements while for a {Mn<sup>III</sup><sub>6</sub>} system, the matrix contains (5<sup>6</sup>*5<sup>6</sup> =) 
+244,140,625 elements. Since the calculation of susceptibility requires the exact diagonalisation of the whole matrix, the matrix needs to be stored 
+in the memory and hence the limitation.</p>
+
+<p></p>
+
+    
+</body></html>

[…]
