### 2021-02-12 WatsonGroupTCD/J2suscep#11 @Swetanshu
The wiki has been copied to the documentation directory.
