### 2021-02-12 WatsonGroupTCD/J2suscep#12 @Swetanshu
Swetanshu wrote most of the code. Prof. G. W. Watson and Prof. W. Schmitt have contributed in the development of the theory implemented in the code and have supervised the development of the code. Prof. G. W. Watson also contributed in the code development, and created the makefiles and the tests.
