### 2021-02-12 #2838 @Swetanshu
Once again thank you @mturiansky, @mfherbst and @dpsanders  for your feedback. Apologies for taking this long to get back to you. We have tried to incorporate almost everything that was suggested. Briefly, the workflow has been streamlined and avoids code duplication, and the repository has been restructured for clarity. The wiki pages include installation and contribution instructions, and simpler examples. Automated tests have also been included to test the proper operation of the code. We have responded to each issue separately in the issue tracker to describe how we addressed each issue. I’ll close them if you are happy with the responses.
The paper has also been updated to remove grammatical errors. We summarise the code and provide an example under the heading ‘J2suscep’. We have not created a separate heading for ‘state of the field’ but we explicitly mention in the paper that the purpose served by ej_calc is not covered elsewhere, and we have included suscep to make the code self-contained.
I look forward to hearing back from you!
