### 2021-03-22 WatsonGroupTCD/J2suscep#1 @mturiansky
Okay, the installation with GNU works for me now on my computer. You should delete the makefile.include in the src/ directory. It's a bit messy, and it distracted me from the fact that you provide clean and consistent makefile.include's in the src/arch/ directory.

@mfherbst if the installation works for you, feel free to close this issue.
