### 2021-03-22 WatsonGroupTCD/J2suscep#3 @mturiansky
The structure looks a bit cleaner. I think this can be closed now.
