### 2021-03-22 WatsonGroupTCD/J2suscep#4 @mturiansky
@dpsanders The contributing page in the wiki with the "for developers" sections looks like it covers the points. Does JOSS want an explicit contributing.md file or is the wiki sufficient?
