### 2021-03-22 WatsonGroupTCD/J2suscep#7 @mturiansky
The script in the bin directory is basically useless. It's not fair to assume J2suscep will be placed in the home directory (or even named as J2suscep).

Why not use command-line arguments to merge the binaries? For example, running "ej_calc -f" would run the formal version, and "ej_calc -s" would be the spin_dens version. Maybe even just one binary for all "j2suscep" with -f and -s above, and -S for suscep or something along those lines.
