### 2021-03-22 WatsonGroupTCD/J2suscep#13 @mturiansky
Consistency
[_review for JOSS_](https://github.com/openjournals/joss-reviews/issues/2838)

I've noticed several spelling errors throughout the output, which should be cleaned up. I also find the lack of consistency between the input file a bit inconvenient to use. For example, "J values" for ej_calc is the number of J values, while for suscep it is the actual list of J values. Why do you need "Hamiltonian Ends" in ej_calc but not in suscep?

On a separate note, is there a way to specify the temperature range or granularity for suscep? That seems like a desirable feature.
