### 2021-03-30 #2838 @mturiansky
Sorry, I should have updated the thread. I checked off a few boxes above, closed some of the issues that were addressed, and opened a new minor issue. @dpsanders I tagged you in one of the issues with a minor question about JOSS's standards. Other than that the code looks to be closer to publishable quality.
