### 2021-04-06 WatsonGroupTCD/J2suscep#7 @mfherbst
I agree with @mturiansky. The script is not useful for me either and I don't see my original points fully addressed in the current version of the code.
