### 2021-04-06 WatsonGroupTCD/J2suscep#8 @mfherbst
I just wanted to reiterate my point of setting up automatic continuous integration, which has not yet been achieved in the repo.

@dpsanders As far as I understand this is not required for JOSS, so I'd consider this optional as far as the review is concerned.
