### 2021-04-06 WatsonGroupTCD/J2suscep#8 @dpsanders
It is correct that continuous integration is not required for JOSS. But I do agree that it is key for modern software construction.
