### 2021-04-06 WatsonGroupTCD/J2suscep#9 @mfherbst
Thanks, that's much better and satisfies me as far as the review is concerned. However, I'd recommend to directly link to the input files from the wiki such that they can be directly downloaded instead of manually searched in the repo.
