### 2021-04-06 WatsonGroupTCD/J2suscep#10 @mfherbst
I see your point, but I fail to see that being stated clearly and explicitly somewhere in the paper.
