### 2021-04-06 WatsonGroupTCD/J2suscep#9 @dpsanders
@Swetanshu: Is there a complete example where I can do the entire calculation with only open-source software?
