### 2021-04-06 WatsonGroupTCD/J2suscep#13 @mfherbst
Related to that: Is it possible to configure the thresholds (e.g. the 1cm^{-1} used to identify generate spin states)? That would probably be handy as well.
