### 2021-04-06 WatsonGroupTCD/J2suscep#6 @mfherbst
Much better now! I'm still not a fan of the large portions of commented code and debug print statements, but I'll leave it here as far as the review is concerned.
