### 2021-04-09 d100812
Minor code update, added temp control to suscep, ej_calc_example update
