### 2021-04-13 aee5e17
Updated the suscep code so that it reads two input files
