+++ src/suscep/suscep.f90
@@ -45,14 +45,15 @@ Program suscep
  Real(kind = real_kind), Dimension(:,:), Allocatable :: spinmat, basis
  ! spinmat stores all possible Ms values for each metal centre
  ! basis stores all possible combinations of Ms values of all metal centres
- Character (Len = 100) :: FileName, read_line, Outfile, inp_file
+ Character (Len = 100) :: read_line, Outfile, inp_file1, inp_file2
  ! inp_file stores the name of the input file
  ! Outfile stores the name of the output file
  Logical :: EndOfFile
- Call Get_Command_Argument(1, FileName)
- inp_file = trim(Filename) !// '.inp'
- Outfile = trim(Filename) // '.out'
- Open (Unit = 11, file = inp_file, action = 'read', position = 'rewind', iostat = ios)
+ Call Get_Command_Argument(1, inp_file1)
+ Call Get_Command_Argument(2, inp_file2)
+ Outfile = trim(inp_file1) // '.out'
+ Open (Unit = 10, file = inp_file1, action = 'read', position = 'rewind', iostat = ios)
+ Open (Unit = 11, file = inp_file2, action = 'read', position = 'rewind', iostat = ios)
  Open (Unit = 12, file = Outfile, status = 'unknown', action = 'readwrite')
  totalspin = 1
  Rewind(11)
@@ -75,6 +76,7 @@ Program suscep
  Call HamilForm(totalspin)
  Deallocate (spin, basis, pseudospin, hamil)
  Deallocate (spinmat, posval, jval, jmatx, jmaty)
+ Close(10)
  Close(11)
  Close(12)
 
@@ -82,7 +84,7 @@ Program suscep
  Contains
 
  Subroutine Init()
-   !!!!! Declaration and initailisation of variables !!!!!
+   !!!!! Declaration and initialisation of variables !!!!!
         Character (Len = 1) :: buff
         Logical :: check, check1
         Integer(kind = int_kind) :: i1, j1
@@ -101,9 +103,9 @@ Program suscep
 
    !!!!! Determining number of magnetic centres (i.e. metal atoms) !!!!!
    !!!!! Initialisation of spin for each metal centre !!!!!
-        Read(11, '(a)', iostat = ios) read_line
+        Read(10, '(a)', iostat = ios) read_line
         If(trim(read_line) .eq. 'magnetic centres') Then
-                Read(11, 168) dmension
+                Read(10, 168) dmension
         Else 
                 Write(*,*) ' Erroneous number of magnetic centres in the input file' 
                 STOP 
@@ -115,6 +117,71 @@ Program suscep
                 spin(i) = 0
         End Do
    !!!!! Determination of magnetic centres and their initialization ends !!!!!
+
+   !!!!! Determining the number of J values, reading them and printing them!!!!!
+        Read(10, '(a)', iostat = ios) read_line
+        If(trim(read_line) .eq. 'No. of J values') then
+                Read(10, 168) no_of_j_val 
+        Else
+                Write(*,*) ' Erroneous No. of J values'
+                STOP 
+        End If
+        Allocate(jval(no_of_j_val))
+        !Read(10, '(a)', iostat = ios) read_line
+   !!!!! J value determination ends !!!!!
+
+      !!!!! Reading in and printing the Spin Hamiltonian !!!!!
+        Read(10, '(a)', iostat = ios) read_line
+        If(trim(read_line) .eq. 'Hamiltonian') Then
+                Read(10, '(a)', iostat = ios) read_line
+        Else
+                Write(*,*) 'No Hamiltonian'
+                STOP
+        End If
+
+
+   !!!!! Determining the number of variables in the Hamiltonian !!!!!
+        i1 = 0
+        Do While (trim(read_line) .ne. 'Hamiltonian Ends')
[…]
