### 2021-04-17 0912ff7
Updated suscep to use 3 input files. Updated j2suscep to execute all 3 codes together
