### 2021-04-17 caccd27
Minor fix for reading jvals in suscep
