### 2021-04-17 b6961c9
Minor fix for reading spins in suscep
