### 2021-04-17 2359e47
Updated test.sh to accommodate new execution procedure for suscep
