### 2021-04-19 64505c3
modified temp loop
