+++ Contributing.md
@@ -1,0 +1,26 @@
+<html><head></head><body>
+
+
+<h1>Contributing</h1>
+<h2>Bugs reports and feature requests</h2>
+
+<p>If you think you've found a bug, please report it on the 
+<a href="https://github.com/WatsonGroupTCD/J2suscep/issues">Issue Tracker</a>.
+You can also propose ideas for new features or ask questions about the design of J2suscep here.
+Poor documentation can also be considered a bug, but please be as specific as possible when asking for improvements.</p>
+
+
+<h2>Code contributions</h2>
+<p>We welcome your contributions towards improving and extending the package and this is managed through Github pull requests.
+For external contributions we prefer the <a href="https://guides.github.com/activities/forking/">"fork and pull"</a>
+workflow while core developers use branches in the main repository:</p>
+
+<blockquote>
+<ol>
+<li>First open an Issue to discuss the proposed contribution. This discussion might include how the changes fit J2suscep's scope and a
+general technical approach.</li>
+<li>Make your own project fork and implement the changes there.</li>
+<li>Open a pull request to merge the changes into the main project. A more detailed discussion can take place there before
+the changes are accepted.</li>
+</ol>
+</blockquote>
