+++ documentation/2c_basic_details.md
@@ -1,12 +1,13 @@
 <html><head></head><body>
 <h1>Basic Details</h1>
 
-<p>This code has been written in FORTRAN 2008. It requires one input (.inp) file that specifies the following:</p>
+<p>This code has been written in FORTRAN 2008. It requires one input file that specifies the following:</p>
 <p>1.	Number of paramagnetic centres in the system.</p>
 <p>2.	The number of unpaired electrons on each centre. </p>
 <p>3.	The coupling behaviour between these centres and the coupling strength.</p>
 <p>4.	The g-value.</p>
 <p>5.	The strength of the applied magnetic field. </p>
+<p>6.	The temperature interval and range in which the value of χ and the χT need to determined. </p>
 <p>This code generates a .out file that contains the value of χ and the χT product at different temperatures. The output 
 also contains details about the size of Hamiltonian matrix and the M<sub>s</sub> values associated with each paramagnetic centre.</p>
 <p></p>
