+++ documentation/1c_basic_details.md
@@ -12,7 +12,7 @@ of the system in the given spin state.</p>
 for the same purpose. This has been done because the value of spin operators are generally not whole numbers and it is difficult to identify singular 
 sets of equations without the use of whole numbers. Since the actual values of spin operators are generally not far off from the nearest whole number, 
 it is highly unlikely that the use of spin operators will remove the singularity. ej_calc_form produces 4 output files – a .out file containing 
-the coupling constants and 3 other files – bigg, nrg_frm_jval_formal and singul. ej_calc_spin produces 2 output files - a .out file containing the coupling constants and nrg_frm_jval. The details about the coupling constants is printed in the .out file. 
+the coupling constants and 3 other files – bigg, nrg_frm_jval_formal and singul. ej_calc_spin produces 2 output files - a .out file containing the coupling constants and nrg_frm_jval_spin. The details about the coupling constants is printed in the .out file. 
 Information contained in the other files is described in the following discussion.</p>
 
 
@@ -26,14 +26,14 @@ the end, only those solutions will be considered valid which lie within 3 standa
 ej_calc_spin reads the ‘singul’ file generated by ej_calc_form and calculates the solutions for only the sets that were found to be non-singular by ej_calc_form.</p>
 
 
-<p>The files nrg_frm_jval_formal and nrg_frm_jval contain the approximate energy of the all possible 2<sup>0.5x</sup> states (x = number of paramagnetic centres). 
+<p>The files nrg_frm_jval_formal and nrg_frm_jval_spin contain the approximate energy of the all possible 2<sup>0.5x</sup> states (x = number of paramagnetic centres). 
 The energy is calculated using the Heisenberg Dirac van Vleck Hamiltonian (H&#770 = -2 &sum;<sub>i>j</sub> J<sub>ij</sub>s<sub>i</sub>s<sub>j</sub>). 
 <font face="Arial, Helvetica, sans-serif" size="-2">[<a href="refs#Heisenberg" class="showTip Heisenberg">Heisenberg</a>], [<a href="refs#Dirac" 
 class="showTip Dirac">Dirac</a>], [<a href="refs#Vanvleck" class="showTip Vanvleck">Vanvleck</a>]</font> 
 The spin operators for all the states are defined by taking the average of the value of spin operator for each paramagnetic centre that was provided as input. 
 Since the variation of the value of spin operators in different spin states is small, the energy obtained this way serves as a reasonable approximation to the 
 actual energy of these states. The true energies of all spin states can, of course, be obtained by diagonalising the full Hamiltonian. Since this diagonalisation 
-becomes computationally quite expensive for large systems, the method employed in this code provides an inexpensive way to obtain a good approximation of the 
+becomes computationally quite expensive for large systems, the method employed in this code provides an inexpensive way to obtain a decent approximation of the 
 energy of these states.</p>
 
 
