+++ documentation/Execution.md
@@ -1,0 +1,211 @@
+<html><head></head><body>
+<h1>Code execution for ej_calc</h1>
+
+<p>The execution needs to be done in two steps.</p>
+<br>
+<p>(i)	The first step is to execute ej_calc_form. If the input files are File1 and File2, then the following command can be used:: </p>
+<pre>ej_calc_form File1 File2</pre>
+<p>It is necessary to run ej_calc_form to produce the singul file which is required to calculate the coupling constant using ej_calc_spin.</p>
+
+<br>
+<p>(ii)	This is followed by the execution of ej_calc_spin as:</p>
+<pre>ej_calc_spin File1 File2</pre>
+<br>
+
+<p>File1 is the file containing the definition of the Hamiltonian.</p>
+<p>File2 is the file containing the value of spin operators for each paramagnetic centre and the energy of different spin states.</p>
+<p>The details about the coupling constants are printed in the files File1_form.out and File1_spin.out for formal spins and spin densities respectively.</p>
+
+
+<h3>Alternative code execution</h3>
+<p>For ease, a bash script (j2suscep) has also provided in the bin directory that can be used to execute the codes. If the input files are File1 and File2, then the code ej_calc_form can be executed as follows:
+</p>
+<pre>j2suscep -f File1 File2</pre>
+<p>The code ej_calc_spin can be executed as follows:</p>
+<pre>j2suscep -s File1 File2</pre>
+<p>The use of the '-s' tag leads to the execution of either ej_calc_spin (if a 'singul' file exists in the working directory), or both ej_calc_form and ej_calc_spin (if a 'singul' file does not exist in the working directory). Therefore, if one is only interested in J-values based on spin densities, the '-s' tag will suffice and there is no need to use the '-f' tag.</p>
+
+
+
+
+
+<p>The details of the format for File1 and File2 are provided in the following discussion.</p>
+<p></p>
+
+
+<br>
+<h2>Format of File1 </h2>
+<p>The format of this file is as follows:</p>
+
+<pre>
+magnetic centres
+&lt;number of paramagnetic centres&gt;
+No. of J values
+&lt;number of j-value>
+Hamiltonian
+&lt;First interaction associated with the first j-value&gt;
+&lt;Second interaction associated with the first j-value&gt;
+&lt;Third interaction associated with the first j-value&gt;
+And so on
+****
+&lt;First interaction associated with the second j-value&gt;
+&lt;Second interaction associated with the second j-value&gt;
+&lt;Third interaction associated with the first j-value&gt;
+And so on
+****
+&lt;First interaction associated with the third j-value&gt;
+&lt;Second interaction associated with the third j-value&gt;
+&lt;Third interaction associated with the first j-value&gt;
+And so on
+Hamiltonian Ends
+
+</pre>
+
+
+<p>The first line has to state 'magnetic centres' and the second line will specify how many magnetic centres are there. 
+Then the number of J-values are specified. After that the Hamiltonian is specified.</p>
+<p>The spin interactions under a single J-value are specified in a single block. 
+The spin interactions for different J-values are separated by '****'. The line 'Hamiltonian Ends' is needed in the end and is not preceded by '****'.</p>
+<p><b>Note:</b> The terms ‘magnetic centres’, ‘No. of J values’, ‘Hamiltonian’ and ‘Hamiltonian Ends’ are case-sensitive.</p>
+
+
+<br>
+<h2>Format of File2</h2>
+<p>The format of this file is as follows:</p>
+
+<pre> 
[…]

[…]
