### 2021-05-13 WatsonGroupTCD/J2suscep#4 @Swetanshu
@mturiansky @dpsanders The contributing.md file has been added to the repo as well now.
