### 2021-05-13 WatsonGroupTCD/J2suscep#7 @Swetanshu
@mfherbst @mturiansky
The script has been modified for better functionality. Once the installation is complete, one needs to either copy/move the contents of the bin folder to a directory in their pathway (the 'PATH' environment variable) or add the bin to the pathway (see wiki). The script also solves the problem of executing multiple executables for ej_calc using the same input file and it can also be used to execute the suscep code. Furthermore, it can be used to execute all the codes sequentially passing the output on one as the input of the next and hence go all the way from the initial input data, through the coupling constants to the susceptibility data output.
The input file for suscep has been broken down into three files in a way such that the input file of ej_calc containing the Hamiltonian can be directly used as one of the three input files that are required. The second file is generated by the script if one tries to execute all codes sequentially. The third input file contains additional information that is needed from the user end and has to be created manually.
One can use the example input files as templates to edit as needed.
