### 2021-05-13 WatsonGroupTCD/J2suscep#8 @Swetanshu
@mfherbst @dpsanders
Continuous integration is not straight forward for Fortran on github. We have included a test system to allow uses to test their installation. We will look at continuous integration in the future.
