### 2021-05-13 WatsonGroupTCD/J2suscep#9 @Swetanshu
@mfherbst  The input files are linked to the wiki now.
@dpsanders  We are aware that open source softwares like ORCA (https://orcaforum.kofo.mpg.de/app.php/portal) can perform such calculations but we don’t have much experience working with it. In principle you should be able to run the same example with ORCA and obtain a very similar or identical solution.
