### 2021-05-13 WatsonGroupTCD/J2suscep#10 @Swetanshu
@mfherbst
We now mention the following in the paper:

“To remove this dependency, more states can be modelled but this procedure results in a large number of solutions, that have to be averaged in some sensible way. Additionally, the problem of singular solutions also arises and such solutions need to be identified and removed. At the moment however, there is no convenient way to do so. The purpose of this package is to provide a means to accomplish this arduous task. Additionally, this package can also calculate the temperature dependence of magnetic susceptibility thereby enabling the comparison of the computational data to experiment. Although other codes like MAGPACK (Ramos et al., 2010) and PHI (Chilton et al., 2013) are available for the calculation of the mangnetic susceptibility, this package provides a one stop solution for the calculation of coupling constants and the determination of the magnetic susceptibility using these coupling constants.” (page 1, line 22 onwards).

Let me know if you want me to mention anything else.
