+++ documentation/Execution.md
@@ -3,7 +3,7 @@
 
 <p>The execution needs to be done in two steps.</p>
 <br>
-<p>(i)	The first step is to execute ej_calc_form. If the input files are File1 and File2, then the following command can be used:: </p>
+<p>(i)	The first step is to execute ej_calc_form. If the input files are <a href="#format-of-file1">File1</a> and <a href="#format-of-file2">File2</a>, then the following command can be used:: </p>
 <pre>ej_calc_form File1 File2</pre>
 <p>It is necessary to run ej_calc_form to produce the singul file which is required to calculate the coupling constant using ej_calc_spin.</p>
 
@@ -12,30 +12,68 @@
 <pre>ej_calc_spin File1 File2</pre>
 <br>
 
-<p>File1 is the file containing the definition of the Hamiltonian.</p>
-<p>File2 is the file containing the value of spin operators for each paramagnetic centre and the energy of different spin states.</p>
+<p><a href="#format-of-file1">File1</a> contains the definition of the Hamiltonian.</p>
+<p><a href="#format-of-file2">File2</a> contains the value of spin operators for each paramagnetic centre and the energy of different spin states.</p>
 <p>The details about the coupling constants are printed in the files File1_form.out and File1_spin.out for formal spins and spin densities respectively.</p>
 
-
-<h3>Alternative code execution</h3>
-<p>For ease, a bash script (j2suscep) has also provided in the bin directory that can be used to execute the codes. If the input files are File1 and File2, then the code ej_calc_form can be executed as follows:
-</p>
-<pre>j2suscep -f File1 File2</pre>
-<p>The code ej_calc_spin can be executed as follows:</p>
-<pre>j2suscep -s File1 File2</pre>
-<p>The use of the '-s' tag leads to the execution of either ej_calc_spin (if a 'singul' file exists in the working directory), or both ej_calc_form and ej_calc_spin (if a 'singul' file does not exist in the working directory). Therefore, if one is only interested in J-values based on spin densities, the '-s' tag will suffice and there is no need to use the '-f' tag.</p>
-
-
-
-
-
-<p>The details of the format for File1 and File2 are provided in the following discussion.</p>
+<p>The details of the format for <a href="#format-of-file1">File1</a> and <a href="#format-of-file2">File2</a> are provided later in this section.</p>
 <p></p>
 
+<h1>Code execution for suscep</h1>
 
-<br>
-<h2>Format of File1 </h2>
-<p>The format of this file is as follows:</p>
+<p>The execution of the suscep code requires three input files, and for input files named <a href="#format-of-file1">File1</a>, <a href="#format-of-file3">File3</a> and <a href="#format-of-file4">File4</a>, the suscep code can be executed in the following way:</p>
+<pre>suscep File1 File3 File4</pre>
+<p><a href="#format-of-file1">File1</a> contains the definition of the Hamiltonian.</p>
+<p><a href="#format-of-file3">File3</a> contains the additional information other than the J-values required to calculate the temperature dependence of magnetic susceptibility.</p>
+<p><a href="#format-of-file4">File4</a> contains the magnitude of J-values to be used for the determination of magnetic susceptibility.</p>
+<p>The temperature dependence of magnetic susceptibility is printed in the file File1.out.</p>
+<p>The details of the format for the input files are provided later in this section.</p>
+
+
+
+<h1>Code execution using script</h1>
+<p>For ease, a bash script (j2suscep) has also provided in the bin directory that can be used to execute the codes separately or together as described below:</p>
+<h2> EJ_Calc <h2>
+
[…]
