### 2021-05-13 WatsonGroupTCD/J2suscep#13 @Swetanshu
@mfherbst @mturiansky
The spelling errors have been removed.
The use of “J values” and “Hamiltonian Ends” has been made consistent now
The temperature range can be specified now in the input file.
Since the energy of different spin states are approximated here to begin with, one may determine the ground spin state and the spin of the highest energy state but it is not advisable to estimate degeneracies from this data. The threshold of 1 cm-1 was arbitrarily chosen to provide some spacing between the rows but should not be interpreted as the definition of degeneracy. To look at the degeneracies properly, one needs to diagonalise the entire matrix of Ms elements.
