### 2021-05-13 WatsonGroupTCD/J2suscep#7 @mturiansky
This seems like an improvement. It would be nice if the script would print out usage information when invalid input is entered instead of just exiting with no message. (Just need to add an `else` block to the script.)
