### 2021-05-21 WatsonGroupTCD/J2suscep#7 @mfherbst
I agree. In a similar vein I would consider adding a `-h` flag or similar, which prints usage.
