### 2021-05-21 WatsonGroupTCD/J2suscep#10 @mfherbst
I'm ok with the current version.

Notice however the typo "mangnetic susceptibility".
