### 2021-05-21 WatsonGroupTCD/J2suscep#9 @mfherbst
@Swetanshu Great. Thanks for adding the links and files. I'm happy now.

Just a short comment regarding @dpsanders question: ORCA is *not* open-source. It is free for academic use only.

Regarding open-source software, my first stop would be Psi4 or pyscf. I have never modelled these systems, so I might be missing something, but should these not be sufficient as well?
