### 2021-05-30 WatsonGroupTCD/J2suscep#7 @Swetanshu
@mturiansky @mfherbst The script has been modified such that it prints out usage information when invalid input is provided or when –h tag is used.
