### 2021-05-30 WatsonGroupTCD/J2suscep#9 @Swetanshu
@dpsanders @mfherbst Based on a quick search, it seems that the both Psi4 (https://psicode.org/psi4manual/master/scf.html#broken-symmetry) and pyscf (https://github.com/pyscf/pyscf/issues/538) can perform broken symmetry calculations and thus, can be used to calculate the coupling constants. However, we have no experience working with these codes.
