### 2021-05-30 #2838 @Swetanshu
Once again thank you @mturiansky, @mfherbst and @dpsanders for your feedback. Now, the usage information for the script is printed when –h tag is used or when invalid input is provided.
@mfherbst The typo in the paper has been removed.
We have responded to the issues separately in the issue tracker as well. I’ll close them if you are happy with the responses.
@dpsanders please let us know if you are happy with the responses to the issues that you opened.
I look forward to hearing back from you.
