### 2021-06-02 WatsonGroupTCD/J2suscep#7 @mfherbst
It's a bit repetitive (for the error section and the `-h` section of the script). Perhaps you can refactor that. Otherwise I'm good.
