### 2021-06-05 WatsonGroupTCD/J2suscep#7 @Swetanshu
@mfherbst The script has been updated to avoid repetition now.
