### 2021-06-16 WatsonGroupTCD/J2suscep#9 @dpsanders
@Swetanshu: Thanks for providing the worked example with Gaussian and the input files. Could you also add the instructions for actually running Gaussian with those input files, how to get the output from Gaussian and take that output and pass it to your tools? In other words, a *complete* worked example.
