### 2021-06-16 WatsonGroupTCD/J2suscep#9 @dpsanders
I feel like the worked example should not be in an appendix, but rather just another section in the documentation, called "Worked example" or similar.
