### 2021-06-16 #2838 @dpsanders
@Swetanshu: I have sent a PR for the JOSS paper with minor edits.

I have also requested that the worked example be extended to show how to run Gaussian, get the output and pass it to your tools to make a complete worked example.
