### 2021-06-22 WatsonGroupTCD/J2suscep#9 @Swetanshu
@dpsanders Details on running Gaussian and obtaining the necessary data has been added to the example. The example has also been moved to a section called "worked example".
