+++ documentation/worked_example.md
@@ -1,6 +1,21 @@
 <html><head></head><body>
 
-<h1>Modelling Broken Symmetry States</h1>
+<h1>Worked Example</h1>
+
+<p>This section describes how one can obtain the necessary data required to prepare <a href="https://github.com/WatsonGroupTCD/J2suscep/wiki/Execution#format-of-file2"> File2</a> and execute J2suscep using the available computational chemistry code <a href="https://gaussian.com/">Gaussian</a>. The example here corresponds to the system described as example1 under 'Example Input Files' subsection of <A HREF="1e_example">EJ_Calc</A> and <A HREF="2e_example">Suscep</A> in this manual. The procedure can be summarised as follows:</p>
+
+<blockquote>
+<ol>
+<li>Model the ferromagnetic and broken symmetry states of the molecule under investigation.</li>
+<li>Extract the spin density data and the energy of the system from the generated output files.</li>
+<li>Use the spin density and energy data to prepare <a href="https://github.com/WatsonGroupTCD/J2suscep/wiki/Execution#format-of-file2"> File2</a> for J2suscep (EJ_calc).</li>
+<li>Run J2suscep.</li>
+</ol>
+</blockquote>
+
+<p><b>Note:</b> It is being assumed that the paths to the utilities and codes described in this example are already added to the PATH variable.</p>
+
+<h2>Modelling Ferromagnetic and Broken Symmetry States</h2>
 
 
 <p>The determination of coupling constants generally relies on the use of broken symmetry states which essentially provide a means to represent 
@@ -104,6 +119,11 @@ and the various fragments. This is followed by the description of the atomic coo
 and the basis set information.</p>
 <p>The complete <a href="https://github.com/WatsonGroupTCD/J2suscep/blob/master/examples/gaussian_files/ferromagnetic.com">input</a> file is provided as part of the repository.</p>
 <br>
+<p>To run the calculation using gaussian09, the following command can be used (assuming that the input file name is ferromagnetic.com):</p>
+<pre>g09 < ferromagnetic.com > ferromagnetic.out</pre>
+<p>This will generate a file called 'ferromagnetic.out' which will contain information in human readable format. Another file called 'ferromagnetic.chk' containing structure and wavefunction information in binary form will also be generated. </p>
+
+
 
 <h3>Antiferromagnetic state</h3>
 
@@ -139,7 +159,10 @@ SDDALL
 
 <p>The term ‘guess=(only,fragment=3)’ on the second line sets up the guess calculation with three fragments 
 (which is the total number of fragments specified in this system). The term ‘pop=minimal’ controls the amount of data 
-written in the log file and is not mandatory for the proper execution of this calculation.</p>
+written in the log file and is not mandatory for the proper execution of this calculation. The execution (assuming the input file name to be 'antifer_guess.com') is performed in the same manner as for the ferromagnetic state:</p>
+<pre>g09 < antifer_guess.com > antifer_guess.out</pre>
+
+
 <p>Once the guess wavefunction is generated, it is analysed for any instability using the following <a href="https://github.com/WatsonGroupTCD/J2suscep/blob/master/examples/gaussian_files/antifer_stab.com">input</a> file:</p>
 <pre>
 %nprocshared=16
@@ -162,6 +185,8 @@ SDDALL
 <p>This calculation reads information from the checkpoint file antiferro.chk generated by the guess calculation. 
 The term ‘stable=opt’ on the fourth line specifies that this calculation checks for the stability of the wavefunction and 
 optimise it if any instability is found. The calculation updates the information in the checkpoint file antiferro.chk.</p>
[…]
