### 2021-06-23 #2838 @dpsanders
👋 @Swetanshu: Once you have done that, please archive the repository at Zenodo or similar, making sure that the title and authors arer exactly the same as in the JOSS paper.

Then please report here the software version and the DOI of the repository.

Thanks!
