+++ JOSS_paper/paper.bib
@@ -30,7 +30,7 @@ title = {Broken Symmetry States of Iron—Sulfur Clusters},
 booktitle = {Encyclopedia of Inorganic Chemistry},
 chapter = {},
 pages = {},
-doi = {https://doi.org/10.1002/0470862106.ia618},
+doi = {10.1002/0470862106.ia618},
 url = {https://onlinelibrary.wiley.com/doi/abs/10.1002/0470862106.ia618},
 eprint = {https://onlinelibrary.wiley.com/doi/pdf/10.1002/0470862106.ia618},
 year = {2009},
