+++ JOSS_paper/paper.bib
@@ -113,7 +113,7 @@ url="https://doi.org/10.1007/128_2015_645"
 
 @Article{Rajeshkumar,
   author  = {Rajeshkumar, Thayalan and Annadata, Harshini V. and Evangelisti, Marco and Langley, Stuart K. and Chilton, Nicholas F. and Murray, Keith S. and Rajaraman, Gopalan},
-  title   = {Theoretical Studies on Polynuclear {$Cu^II_5Gd^I ^I ^I_n$} Clusters (n = 4, 2): Towards Understanding Their Large Magnetocaloric Effect},
+  title   = {Theoretical Studies on Polynuclear {$Cu^II_5$$Gd^III_n$} Clusters (n = 4, 2): Towards Understanding Their Large Magnetocaloric Effect},
   doi     = {10.1021/ic502651w},
   eprint  = {https://doi.org/10.1021/ic502651w},
   note    = {PMID: 25615271},
