### 2021-06-24 #2838 @Swetanshu
@dpsanders The software has been archived on zenodo. The doi is 10.5281/zenodo.5024912
The version is 1.0.
Let me know if there is anything else that needs to be done from my end.
Also, can you let me know if I can close the issues that are still open?

Thanks again!
